backend rewrite

+++ code/Server.Api/Controllers/ChannelController.cs
@@ -1,10 +1,10 @@
-using System.Threading.Tasks;
-using System.Collections.Generic;
-using Microsoft.AspNetCore.Mvc;
+using System.Linq;
 using Server.Api.Dtos;
 using Server.Api.Models;
+using System.Threading.Tasks;
 using Server.Api.Repositories;
-using System.Linq;
+using Microsoft.AspNetCore.Mvc;
+using System.Collections.Generic;
 using Microsoft.AspNetCore.Authorization;
 
 namespace Server.Api.Controllers
@@ -13,17 +13,16 @@ namespace Server.Api.Controllers
     [Route("[controller]")]
     public class ChannelController : ControllerBase
     {
-        private readonly IChannelRepository _channelRepository;
         private readonly ICourseRepository _courseRepository;
+        private readonly IChannelRepository _channelRepository;
 
         public ChannelController(IChannelRepository channelRepository, ICourseRepository courseRepository)
         {
-            _channelRepository = channelRepository;
             _courseRepository = courseRepository;
+            _channelRepository = channelRepository;
         }
 
         [Authorize(Roles = "student,prof")]
-
         [HttpGet]
         public async Task<ActionResult<IEnumerable<GetPartialChannelDto>>> GetChannels()
         {
@@ -40,7 +39,6 @@ namespace Server.Api.Controllers
         }
 
         [Authorize(Roles = "student,prof")]
-
         [HttpGet("{id}")]
         public async Task<ActionResult<GetChannelDto>> GetChannel(int id)
         {
@@ -52,15 +50,13 @@ namespace Server.Api.Controllers
         }
 
         [Authorize(Roles = "prof")]
-
         [HttpPost]
         public async Task<ActionResult<GetPartialChannelDto>> CreateChannel(CreateChannelDto dto)
         {
-            TextChannel channel = new()
-            {
+            TextChannel channel = new() {
                 name = dto.name,
-                course = await _courseRepository.GetAsync(dto.courseId),
-                messages = new List<Message>()
+                messages = new List<Message>(),
+                course = await _courseRepository.GetAsync(dto.courseId)
             };
 
             await _channelRepository.CreateAsync(channel);
@@ -68,21 +64,20 @@ namespace Server.Api.Controllers
         }
 
         [Authorize(Roles = "prof")]
-
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteChannel(int id)
         {
+            var existing = _channelRepository.GetAsync(id);
+            if (existing == null) return NotFound();
             await _channelRepository.DeleteAsync(id);
             return NoContent();
         }
 
         [Authorize(Roles = "prof")]
-
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateChannel(int id, CreateChannelDto dto)
         {
-            TextChannel channel = new()
-            {
+            TextChannel channel = new() {
                 id = id,
                 name = dto.name,
                 course = await _courseRepository.GetAsync(dto.courseId)
