+++ code/Server.Api/Controllers/FieldStudyController.cs
@@ -1,13 +1,10 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Server.Api.Models;
-using Server.Api.Repositories;
 using Server.Api.Dtos;
+using Server.Api.Models;
+using System.Threading.Tasks;
+using Server.Api.Repositories;
+using Microsoft.AspNetCore.Mvc;
+using System.Collections.Generic;
 using Microsoft.AspNetCore.Identity;
-
-using System;
-using System.ComponentModel;
 using Microsoft.AspNetCore.Authorization;
 
 namespace Server.Api.Controllers
@@ -16,36 +13,33 @@ namespace Server.Api.Controllers
     [Route("[controller]")]
     public class FieldOfStudyController : ControllerBase
     {
-        private readonly IFieldOfStudyRepository _fieldOfStudyRepository;
         private readonly UserManager<User> _userManager;
+        private readonly IFieldOfStudyRepository _fieldOfStudyRepository;
 
         public FieldOfStudyController(IFieldOfStudyRepository fieldOfStudyRepository, UserManager<User> userManager)
         {
-            _fieldOfStudyRepository = fieldOfStudyRepository;
             _userManager = userManager;
+            _fieldOfStudyRepository = fieldOfStudyRepository;
         }
-        [Authorize(Roles = "student,prof")]
 
+        [Authorize(Roles = "student,prof")]
         [HttpGet]
         public async Task<ActionResult<IEnumerable<FieldOfStudy>>> GetFieldOfStudies()
         {
             var fieldOfStudies = await _fieldOfStudyRepository.GetAllAsync();
             return Ok(fieldOfStudies);
         }
-        [Authorize(Roles = "student,prof")]
 
+        [Authorize(Roles = "student,prof")]
         [HttpGet("{id}")]
         public async Task<ActionResult<FieldOfStudy>> GetFieldOfStudy(int id)
         {
             var fieldOfStudy = await _fieldOfStudyRepository.GetAsync(id);
-            if (fieldOfStudy == null)
-            {
-                return NotFound();
-            }
+            if (fieldOfStudy == null) return NotFound();
             return Ok(fieldOfStudy);
         }
-        [Authorize(Roles = "student,prof")]
 
+        [Authorize(Roles = "student,prof")]
         [HttpGet("user")]
         public async Task<ActionResult<FieldOfStudy>> GetUserFieldOfStudy([FromQuery] string email)
         {
@@ -56,41 +50,42 @@ namespace Server.Api.Controllers
             if (fieldOfStudy == null) return NotFound();
             return Ok(fieldOfStudy);
         }
-        [Authorize(Roles = "prof")]
 
+        [Authorize(Roles = "prof")]
         [HttpPost]
         public async Task<ActionResult> CreateFieldOfStudy(CreateFieldOfStudyDto createFieldOfStudyDto)
         {
-            FieldOfStudy fieldOfStudy = new()
-            {
+            FieldOfStudy fieldOfStudy = new() {
                 name = createFieldOfStudyDto.name,
                 fullName = createFieldOfStudyDto.fullName,
                 isBachelor = createFieldOfStudyDto.isBachelor
             };
+
             await _fieldOfStudyRepository.CreateAsync(fieldOfStudy);
             return Ok();
         }
 
         [Authorize(Roles = "prof")]
-
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteFieldOfStudy(int id)
         {
+            var existing = _fieldOfStudyRepository.GetAsync(id);
+            if (existing == null) return NotFound();
             await _fieldOfStudyRepository.DeleteAsync(id);
             return Ok();
         }
-        [Authorize(Roles = "prof")]
 
+        [Authorize(Roles = "prof")]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateFieldOfStudy(int id, FieldOfStudyDto updateFieldOfStudyDto)
         {
-            FieldOfStudy fieldOfStudy = new()
-            {
+            FieldOfStudy fieldOfStudy = new() {
                 id = id,
                 name = updateFieldOfStudyDto.name,
                 fullName = updateFieldOfStudyDto.fullName,
                 isBachelor = updateFieldOfStudyDto.isBachelor
             };
+
             await _fieldOfStudyRepository.UpdateAsync(fieldOfStudy);
             return Ok();
         }
