+++ code/Server.Api/Controllers/TopicController.cs
@@ -1,10 +1,10 @@
 using System.Linq;
-using System.Threading.Tasks;
-using System.Collections.Generic;
-using Microsoft.AspNetCore.Mvc;
 using Server.Api.Dtos;
 using Server.Api.Models;
+using System.Threading.Tasks;
 using Server.Api.Repositories;
+using Microsoft.AspNetCore.Mvc;
+using System.Collections.Generic;
 using Microsoft.AspNetCore.Authorization;
 
 namespace Server.Api.Controllers
@@ -18,17 +18,15 @@ namespace Server.Api.Controllers
 
         private static GetTopicDto ToDto(Topic topic)
         {
-            return new GetTopicDto
-            {
+            return new GetTopicDto {
                 id = topic.id,
                 name = topic.name,
-                course = new GetPartialCourseDto
-                {
+                questions = topic.questions.Select(question => GetPartialQuestionDto.Convert(question)).ToList(),
+                course = new GetPartialCourseDto {
                     id = topic.course.id,
                     name = topic.course.name,
-                    number = topic.course.number,
-                },
-                questions = topic.questions.Select(question => GetPartialQuestionDto.Convert(question)).ToList()
+                    number = topic.course.number
+                }
             };
         }
 
@@ -39,37 +37,32 @@ namespace Server.Api.Controllers
         }
 
         [Authorize(Roles = "student,prof")]
-
         [HttpGet]
         public async Task<ActionResult<IEnumerable<GetTopicDto>>> GetTopics()
         {
             var topics = await _topicRepository.GetAllAsync();
             return Ok(topics.Select(topic => ToDto(topic)));
         }
-        [Authorize(Roles = "student,prof")]
 
+        [Authorize(Roles = "student,prof")]
         [HttpGet("{id}")]
         public async Task<ActionResult<GetTopicDto>> GetTopic(int id)
         {
             var topic = await _topicRepository.GetAsync(id);
-            if (topic == null)
-            {
-                return NotFound();
-            }
+            if (topic == null) return NotFound();
             return Ok(ToDto(topic));
         }
 
         [Authorize(Roles = "prof")]
-
         [HttpPost]
         public async Task<ActionResult<Topic>> CreateTopic(CreateTopicDto dto)
         {
-            Topic topic = new()
-            {
+            Topic topic = new() {
                 name = dto.name,
                 course = await _courseRepository.GetAsync(dto.courseId),
                 questions = new List<Question>()
             };
+
             await _topicRepository.CreateAsync(topic);
             return Ok(topic);
         }
@@ -79,21 +72,22 @@ namespace Server.Api.Controllers
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteTopic(int id)
         {
+            var existing = _topicRepository.GetAsync(id);
+            if (existing == null) return NotFound();
             await _topicRepository.DeleteAsync(id);
             return NoContent();
         }
 
         [Authorize(Roles = "prof")]
-
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateTopic(int id, CreateTopicDto dto)
         {
-            Topic topic = new()
-            {
+            Topic topic = new() {
                 id = id,
                 name = dto.name,
                 course = await _courseRepository.GetAsync(dto.courseId)
             };
+
             await _topicRepository.UpdateAsync(topic);
             return NoContent();
         }
